decouple logic of new node into separate logic
